Oops, fix test I broke with #12.

+++ test/lib/SDL_mixer.jl
@@ -5,8 +5,8 @@ SDL2_pkg_dir = joinpath(Pkg.dir(), "SDL2")
 audio_example_assets_dir = joinpath(SDL2_pkg_dir, "src/examples/audio_example")
 
 # check that an audio device if available
-SDL2.Init(Int32(SDL2.INIT_VIDEO))
-device = SDL2.Mix_OpenAudio(Int32(22050), Int32(SDL2.MIX_DEFAULT_FORMAT), Int32(2), Int32(1024) )
+SDL2.Init(UInt32(SDL2.INIT_VIDEO))
+device = SDL2.Mix_OpenAudio(Int32(22050), UInt16(SDL2.MIX_DEFAULT_FORMAT), Int32(2), Int32(1024) )
 
 if device == 0
 SDL2.Mix_CloseAudio()
